basic-setup-with-dags/requirements.txt
apache-airflow[statsd]==3.0.0
airflow-provider-duckdb==0.1.0
apache-airflow-providers-postgres==5.13.1
apache-airflow-providers-common-sql==1.26.0
apache-airflow-providers-fab[common.compat]==2.0.2
apache-airflow-providers-apache-kafka==1.9.0

astronomer-cosmos==1.10.1

duckdb==1.2.0
authlib==1.5.2
python-ldap==3.4.4
dbt-postgres==1.9.0
asgiref==3.8.1

beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
frozendict==2.4.6
html5lib==1.1
idna==3.10
lxml==5.3.0
multitasking==0.0.11
mysql-connector-python==9.1.0
numpy==2.2.1
pandas==2.2.3
peewee==3.17.8
platformdirs==4.3.6
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.3.0
webencodings==0.5.1
yfinance==0.2.51
pyodbc>=5.2.0
requests>=2.32.4